requirements.txt
valyu
ollama
python-dotenv
matplotlib
seaborn
tqdm
pandas
tiktoken
